Adding protocols for plane type operations
